Created hooks folter, indent useDevice with 2 spaces
